sync word can be 64 bits

+++ gr-capture_tools/lib/sync_and_strobe_impl.h
@@ -52,10 +52,10 @@ namespace gr {
         float d_extreme_val_high;
         float d_extreme_val_low;
         float d_frozen_avg_val;
-        uint32_t d_input_buffer;
+        uint64_t d_input_buffer;
 
-        uint32_t d_sync_word;
-        uint32_t d_sync_word_mask;
+        uint64_t d_sync_word;
+        uint64_t d_sync_word_mask;
         int d_sync_word_len;
         int d_start_counter;
         int d_packet_counter;
